requirements-ci.txt
black==22.10.0

flake8==5.0.4
flake8-bugbear==22.10.27
flake8-docstrings==1.6.0
flake8-eradicate==1.4.0

isort==5.10.1
mypy==0.991
pep8-naming==0.13.2

pytest==7.2.0
pytest-cov==4.0.0
pytest-flask==1.2.0
pytest-xdist==3.0.2

types-flask-cors==3.0.10
